// src/app/shared/enums/PlaceType.ts
export enum PlaceTypeCityEnum {
  ADEGA = 'ADEGA',
  BAR = 'BAR',
  RESTAURANTE = 'RESTAURANTE',
  CAFETERIA = 'CAFETERIA',
  CHURRASCARIA = 'CHURRASCARIA',
  DOCERIA = 'DOCERIA',
  EMPORIO = 'EMPORIO',
  HAMBURGUERIA = 'HAMBURGUERIA',
  MERCADO = 'MERCADO',
  PIZZARIA = 'PIZZARIA',
  SORVETERIA = 'SORVETERIA',
  TABACARIA = 'TABACARIA'
}

export enum PlaceTypeBeachEnum {
  QUIOSQUE = 'QUIOSQUE',
  CARRINHO = 'CARRINHO'
}